create mergeRollupMiddleware, simplify CSV import, will add to genbank import
